+++ plugin/plugin.py
@@ -33,8 +33,8 @@ from __init__ import _
 config.OpenWebif = ConfigSubsection()
 config.OpenWebif.enabled = ConfigYesNo(default=True)
 config.OpenWebif.identifier = ConfigYesNo(default=True)
-config.OpenWebif.identifier.custom = ConfigYesNo(default=False)
-config.OpenWebif.identifier.text = ConfigText(default = "", fixed_size = False)
+config.OpenWebif.identifier_custom = ConfigYesNo(default=False)
+config.OpenWebif.identifier_text = ConfigText(default = "", fixed_size = False)
 config.OpenWebif.port = ConfigInteger(default = 80, limits=(1, 65535) )
 config.OpenWebif.streamport = ConfigInteger(default = 8001, limits=(1, 65535) )
 config.OpenWebif.auth = ConfigYesNo(default=False)
@@ -44,6 +44,7 @@ config.OpenWebif.webcache = ConfigSubsection()
 config.OpenWebif.webcache.collapsedmenus = ConfigText(default = "remote", fixed_size = False)
 config.OpenWebif.webcache.remotegrabscreenshot = ConfigYesNo(default = True)
 config.OpenWebif.webcache.zapstream = ConfigYesNo(default = False)
+config.OpenWebif.webcache.epg_desc_search = ConfigYesNo(default = False)
 # HTTPS
 config.OpenWebif.https_enabled = ConfigYesNo(default=False)
 config.OpenWebif.https_port = ConfigInteger(default = 443, limits=(1, 65535) )
@@ -69,7 +70,7 @@ config.OpenWebif.epg_encoding = ConfigSelection(default = 'utf-8', choices = [ '
 										'iso-8859-9',
 										'iso-8859-10',
 										'iso-8859-16'])
-
+imagedistro = getInfo()['imagedistro']
 
 class OpenWebifConfig(Screen, ConfigListScreen):
 	skin = """
@@ -108,9 +109,9 @@ class OpenWebifConfig(Screen, ConfigListScreen):
 		if config.OpenWebif.enabled.value:
 			self.list.append(getConfigListEntry(_("Show box name in header"), config.OpenWebif.identifier))
 			if config.OpenWebif.identifier.value:
-				self.list.append(getConfigListEntry(_("Use custom box name"), config.OpenWebif.identifier.custom))
-				if config.OpenWebif.identifier.custom.value:
-					self.list.append(getConfigListEntry(_("Custom box name"), config.OpenWebif.identifier.text))
+				self.list.append(getConfigListEntry(_("Use custom box name"), config.OpenWebif.identifier_custom))
+				if config.OpenWebif.identifier_custom.value:
+					self.list.append(getConfigListEntry(_("Custom box name"), config.OpenWebif.identifier_text))
 			self.list.append(getConfigListEntry(_("HTTP port"), config.OpenWebif.port))
 			self.list.append(getConfigListEntry(_("Enable HTTP Authentication"), config.OpenWebif.auth))
 			self.list.append(getConfigListEntry(_("Enable HTTPS"), config.OpenWebif.https_enabled))
@@ -124,7 +125,8 @@ class OpenWebifConfig(Screen, ConfigListScreen):
 			self.list.append(getConfigListEntry(_("Smart services renaming for XBMC"), config.OpenWebif.xbmcservices))
 			self.list.append(getConfigListEntry(_("Enable Parental Control"), config.OpenWebif.parentalenabled))
 			self.list.append(getConfigListEntry(_("Add service name to stream information"), config.OpenWebif.service_name_for_stream))
-			self.list.append(getConfigListEntry(_("Character encoding for EPG data"), config.OpenWebif.epg_encoding))
+			if imagedistro in ("VTi-Team Image"):
+				self.list.append(getConfigListEntry(_("Character encoding for EPG data"), config.OpenWebif.epg_encoding))
 
 		self["config"].list = self.list
 		self["config"].l.setList(self.list)
@@ -178,7 +180,7 @@ def startSession(reason, session):
 
 def main_menu(menuid, **kwargs):
 	if menuid == "network":
-		return [("OpenWebif", confplug, "openwebif", 37)]
+		return [("OpenWebif", confplug, "openwebif", 45)]
 	else:
 		return []
 
@@ -187,7 +189,6 @@ def Plugins(**kwargs):
 		PluginDescriptor(where=[PluginDescriptor.WHERE_SESSIONSTART], fnc=startSession),
 		PluginDescriptor(where=[PluginDescriptor.WHERE_NETWORKCONFIG_READ], fnc=IfUpIfDown),
 		]
-	imagedistro = getInfo()['imagedistro']
 	screenwidth = getDesktop(0).size().width()
 	if imagedistro in ("openatv"):
 		result.append(PluginDescriptor(name="OpenWebif", description=_("OpenWebif Configuration"), where = PluginDescriptor.WHERE_MENU, fnc = main_menu))