+++ CHANGES.md
@@ -3,9 +3,28 @@
 **FEATURES**
 * powertimer api
 
-## Version 0.4.6
+## Version 0.4.8
 ## (in development)
 **FEATURES**
+* get movie text info
+* improve movie trash
+* improve timers for radio
+* deprecated -> "Smart services renaming for XBMC" use url parameter "renameserviceforxmbc=1" for getallservices instead 
+
+## Version 0.4.7
+**FEATURES**
+* improve edit movie tags
+
+## Version 0.4.6
+**FEATURES**
+* add more supported box models
+
+**FIXES**
+* translations
+* encoding problems
+* owbranding
+* streaming problems
+* parental control
 
 ## Version 0.4.5
 **FEATURES**